Code cleanup and reorganization
