+++ benches/bench.rs
@@ -1,0 +1,4 @@
+#![feature(test)]
+#[cfg(test)]
+
+mod magic_benches;
