+++ src/bit_twiddles.rs
@@ -1,5 +1,3 @@
-use test::Bencher;
-use test;
 
 static POPCNT8: &'static [u8] = &[
     0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 4,
@@ -172,128 +170,4 @@ fn popcount_old(x: u64) -> u8 {
         POPCNT8[(x >> 16) & 0xFF] +
         POPCNT8[(x >> 8) & 0xFF] +
         POPCNT8[x & 0xFF]
-}
-
-
-
-
-pub const TRAILS: u64 = 17_000;
-
-
-#[bench]
-fn evs_bench_bitscan_djuie(b: &mut Bencher) {
-    b.iter(|| {
-        let n: u64 = test::black_box(TRAILS);
-        (0..n).fold(0, |a, c| {
-            let mut x: u64 = very_sparse_random(c.wrapping_mul(909090909090909091));
-            if x == 0 { x = 1;} else { x = bit_scan_forward(x) as u64;}
-            a ^ (x) }
-        )
-    })
-}
-
-
-
-#[bench]
-fn evs_bench_popcount_rust(b: &mut Bencher) {
-    b.iter(|| {
-        let n: u64 = test::black_box(TRAILS);
-        (0..n).fold(0, |a, c| {
-            let mut x: u64 = very_sparse_random(c.wrapping_mul(909090909090909091));
-            if x == 0 { x = 1;} else { x = popcount64(x) as u64;}
-            a ^ (x) }
-        )
-    })
-}
-
-#[bench]
-fn evs_bench_popcount_old(b: &mut Bencher) {
-    b.iter(|| {
-        let n: u64 = test::black_box(TRAILS);
-        (0..n).fold(0, |a, c| {
-            let mut x: u64 = very_sparse_random(c.wrapping_mul(909090909090909091));
-            if x == 0 { x = 1;} else { x = popcount_old(x) as u64;}
-            a ^ (x) }
-        )
-    })
-}
-
-
-#[bench]
-fn evs_bench_lsb_pop_rust(b: &mut Bencher) {
-    b.iter(|| {
-        let n: u64 = test::black_box(TRAILS);
-        (0..n).fold(0, |a, c| {
-            let mut x: u64 = very_sparse_random(c.wrapping_mul(909090909090909091));
-            if x == 0 { x = 1;} else { x = lsb(x) as u64;}
-            a ^ (x)
-        })
-    })
-}
-
-
-#[bench]
-fn evs_bench_randomize_super_sparse(b: &mut Bencher) {
-    b.iter(|| {
-        let n: u64 = test::black_box(TRAILS);
-        (0..n).fold(0, |a, c| {
-            let mut x: u64 = very_sparse_random(c.wrapping_mul(909090909090909091));
-            if x == 0 { x = 1;}
-            a ^ (x) }
-        )
-    })
-}
-
-
-
-#[inline]
-fn randomize(x: u64) -> u64{
-    (!random_2(random_1(x))  ^ random_4(!x)) & random_1(x)
-}
-
-// Densest
-#[inline]
-fn randomize2(x: u64) -> u64{
-    random_1(!random_2(x))  ^ !random_3(!x) ^ random_4(x)
-}
-
-#[inline]
-fn randomize_sparse (x: u64) -> u64{
-    (!random_3(random_2(!x))) & random_1(x) & random_4(x)
-}
-
-#[inline]
-fn very_sparse_random (x: u64) -> u64{
-    ((!8512677386048191063) ^ !(x + 6)).wrapping_mul(!x) & ((!1030501117050341) ^ x).wrapping_mul(!x) & (! 1030507050301 as u64).wrapping_mul(!x) & random_2(x) & (!x).wrapping_mul(2685821657736338717) & !random_4(x) & random_3(!x) & (x).wrapping_mul(268582165773633871)
-}
-
-
-
-
-fn random_3(x: u64) -> u64 {
-    let mut c: u64 = x.wrapping_shr(12);
-    c ^= x.wrapping_shl(25);
-    c ^= x.wrapping_shr(27);
-    x.wrapping_mul(1030507050301) ^ c
-}
-
-fn random_2(x: u64) -> u64 {
-    let mut c: u64 = x.wrapping_shr(12);
-    c ^= x.wrapping_shl(25);
-    c ^= x.wrapping_shr(27);
-    x.wrapping_mul(8512677386048191063) ^ c
-}
-
-fn random_1(x: u64) -> u64 {
-    let mut c: u64 = x.wrapping_shr(12);
-    c ^= x.wrapping_shl(25);
-    c ^= x.wrapping_shr(27);
-    x.wrapping_mul(2685821657736338717) ^ c
-}
-
-fn random_4(x: u64) -> u64 {
-    let mut c: u64 = x.wrapping_shr(12);
-    c ^= x.wrapping_shl(25);
-    c ^= x.wrapping_shr(27);
-    x.wrapping_mul(399899999999999 ) ^ c
 }
