+++ src/bots/bot_advanced.rs
@@ -341,23 +341,23 @@ fn bench_bot_ply_3__main_bot(b: &mut Bencher) {
     })
 }
 
-#[bench]
-fn bench_bot_ply_4__main_bot(b: &mut Bencher) {
-    use templates::TEST_FENS;
-    b.iter(|| {
-        let mut b: Board = test::black_box(Board::default());
-        let iter = TEST_FENS.len();
-        let mut i = 0;
-        (0..iter).fold(0, |a: u64, c| {
-            //            println!("{}",TEST_FENS[i]);
-            let mut b: Board = test::black_box(Board::new_from_fen(TEST_FENS[i]));
-            let mov = AdvancedBot::best_move_depth(b.shallow_clone(), &timer::Timer::new(20), 4);
-            b.apply_move(mov);
-            i += 1;
-            a ^ (b.zobrist()) }
-        )
-    })
-}
+//#[bench]
+//fn bench_bot_ply_4__main_bot(b: &mut Bencher) {
+//    use templates::TEST_FENS;
+//    b.iter(|| {
+//        let mut b: Board = test::black_box(Board::default());
+//        let iter = TEST_FENS.len();
+//        let mut i = 0;
+//        (0..iter).fold(0, |a: u64, c| {
+//            //            println!("{}",TEST_FENS[i]);
+//            let mut b: Board = test::black_box(Board::new_from_fen(TEST_FENS[i]));
+//            let mov = AdvancedBot::best_move_depth(b.shallow_clone(), &timer::Timer::new(20), 4);
+//            b.apply_move(mov);
+//            i += 1;
+//            a ^ (b.zobrist()) }
+//        )
+//    })
+//}
 
 
 //#[bench]
