+++ src/bots/bot_jamboree.rs
@@ -7,34 +7,14 @@ use rayon;
 use test::Bencher;
 use test;
 
+use super::BestMove;
+
+
 
 pub struct JamboreeSearcher {
     board: Board,
     timer: Timer,
 }
-
-pub struct BestMove {
-    best_move: Option<BitMove>,
-    score: i16,
-}
-
-impl BestMove {
-    pub fn new(score: i16) -> Self {
-        BestMove{
-            best_move: None,
-            score: score
-        }
-    }
-
-    pub fn negate(mut self) -> Self {
-        self.score *= -1;
-        self
-    }
-
-    pub fn score(&self) -> i16 {self.score}
-}
-
-
 
 
 const MAX_PLY: u16 = 5;
