+++ src/eval.rs
@@ -2,8 +2,6 @@ use board::Board;
 use std::i16;
 use templates::*;
 use bit_twiddles::*;
-use magic_helper::*;
-
 
 lazy_static! {
     pub static ref BISHOP_POS: [[i16; SQ_CNT]; PLAYER_CNT] = [ flatten(flip(BISHOP_POS_ARRAY)), flatten(BISHOP_POS_ARRAY) ];
