+++ src/magic_helper.rs
@@ -3,8 +3,6 @@
 use bit_twiddles::*;
 use templates::*;
 use std::{mem,slice,cmp};
-use test::Bencher;
-use test;
 
 
 const ROOK_M_SIZE: usize = 102_400;
@@ -840,143 +838,42 @@ pub fn diff(x: u8, y: u8) -> u8 {
 
 
 
+#[cfg(test)]
+mod tests {
 
-#[test]
-fn test_king_mask_gen() {
-    let arr = gen_king_moves().to_vec();
-    let sum = arr.iter().fold(0 as  u64,|a, &b| a + (popcount64(b) as u64));
-    assert_eq!(sum, (3*4) + (5 * 6 * 4) + (8 * 6 * 6));
+    use magic_helper::*;
+    use test;
+
+    #[test]
+    fn test_king_mask_gen() {
+        let arr = gen_king_moves().to_vec();
+        let sum = arr.iter().fold(0 as  u64,|a, &b| a + (popcount64(b) as u64));
+        assert_eq!(sum, (3*4) + (5 * 6 * 4) + (8 * 6 * 6));
+    }
+
+    #[test]
+    fn test_knight_mask_gen() {
+        let arr = gen_knight_moves().to_vec();
+        let sum = arr.iter().fold(0 as  u64,|a, &b| a + (popcount64(b) as u64));
+        assert_eq!(sum, (2 * 4) + (4 * 4) + (3 * 2 * 4) + (4 * 4 * 4) + (6 * 4 * 4) + (8 * 4 * 4));
+    }
+
+    #[test]
+    fn occupancy_and_sliding() {
+        let rook_deltas: [i8; 4] = [8,1,-8,-1];
+        assert_eq!(popcount64(sliding_attack(&rook_deltas, 0, 0)),14);
+        assert_eq!(popcount64(sliding_attack(&rook_deltas, 0, 0xFF00)),8);
+        assert_eq!(popcount64(sliding_attack(&rook_deltas, 19, 0)),14);
+    }
+
+    #[test]
+    fn rmagics() {
+        let mstruct = MRookTable::init();
+        assert_eq!(mem::size_of_val(&mstruct), 2584);
+        let bstruct = MBishopTable::init();
+        assert_eq!(mem::size_of_val(&bstruct), 2584);
+    }
+
+
 }
 
-#[test]
-fn test_knight_mask_gen() {
-    let arr = gen_knight_moves().to_vec();
-    let sum = arr.iter().fold(0 as  u64,|a, &b| a + (popcount64(b) as u64));
-    assert_eq!(sum, (2 * 4) + (4 * 4) + (3 * 2 * 4) + (4 * 4 * 4) + (6 * 4 * 4) + (8 * 4 * 4));
-}
-
-#[test]
-fn occupancy_and_sliding() {
-    let rook_deltas: [i8; 4] = [8,1,-8,-1];
-    assert_eq!(popcount64(sliding_attack(&rook_deltas, 0, 0)),14);
-    assert_eq!(popcount64(sliding_attack(&rook_deltas, 0, 0xFF00)),8);
-    assert_eq!(popcount64(sliding_attack(&rook_deltas, 19, 0)),14);
-}
-
-#[test]
-fn rmagics() {
-    let mstruct = MRookTable::init();
-    assert_eq!(mem::size_of_val(&mstruct), 2584);
-    let bstruct = MBishopTable::init();
-    assert_eq!(mem::size_of_val(&bstruct), 2584);
-}
-
-
-#[bench]
-fn bench_rook_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let x: u64 = m.rook_moves(a,c);
-            a ^ (x) }
-        )
-    })
-}
-
-
-#[bench]
-fn bench_bishop_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let x: u64 = m.bishop_moves(a,c);
-            a ^ (x) }
-        )
-    })
-}
-
-#[bench]
-fn bench_queen_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let x: u64 = m.queen_moves(a,c);
-            a ^ (x) }
-        )
-    })
-}
-
-#[bench]
-fn bench_king_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let x: u64 = m.king_moves(c);
-            a ^ (x) }
-        )
-    })
-}
-
-#[bench]
-fn bench_knight_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let x: u64 = m.knight_moves(c);
-            a ^ (x) }
-        )
-    })
-}
-
-// Benefits from locality
-#[bench]
-fn bench_sequential_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let mut x: u64 = m.knight_moves(c);
-            x ^= m.king_moves(c);
-            x ^= m.bishop_moves(x,c);
-            x ^= m.rook_moves(x,c);
-            x ^= m.queen_moves(x,c);
-            a ^ (x) }
-        )
-    })
-}
-
-
-// Stutters so Cache must be refreshed more often
-#[bench]
-fn bench_stutter_lookup(b: &mut Bencher) {
-    let m = MagicHelper::new();
-    b.iter(|| {
-        let n: u8 = test::black_box(64);
-        (0..n).fold(0, |a: u64, c| {
-            let mut x: u64 = m.queen_moves(a,c);
-            x ^= m.king_moves(c);
-            x ^= m.bishop_moves(x,c);
-            x ^= m.knight_moves(c);
-            x ^= m.rook_moves(x,c);
-            a ^ (x) }
-        )
-    })
-}
-
-#[bench]
-// NOTE: This takes a while :/
-fn bench_creation(b: &mut Bencher) {
-    b.iter(|| {
-        let n: u8 = test::black_box(1);
-        (0..n).fold(0, |a: u64, c| {
-            let m = MagicHelper::new();
-            let x: u64 = m.king_moves(c);
-            a ^ (x) }
-        )
-    })
-}
